create useModality hook
